Make test structure reflect changed module structure
